feat(vim): persistent undo
Persistent undo will be very important, considering that I need to
configure and re-open the edtior several times a day. That will be
pretty annoying if I have lost all the undo history (that would never
happen on PyCharm, by the way).

Also "options" -> "opts" as it is already deprecated.